+++ substrate/frame/transaction-storage/src/lib.rs
@@ -534,18 +534,16 @@ pub mod pallet {
 
 		pub fn indexed_transactions(
 			block: BlockNumberFor<T>,
-		) -> alloc::vec::Vec<sp_transaction_storage_proof::runtime_api::IndexedTransactionInfo> {
+		) -> alloc::vec::Vec<sp_transaction_storage_proof::IndexedTransactionInfo> {
 			Transactions::<T>::get(block)
 				.map(|txs| {
 					txs.into_iter()
-						.map(|tx| {
-							sp_transaction_storage_proof::runtime_api::IndexedTransactionInfo {
+						.map(|tx| sp_transaction_storage_proof::IndexedTransactionInfo {
 							content_hash: tx.content_hash.into(),
 							size: tx.size,
-							hashing: sp_transaction_storage_proof::runtime_api::HashingAlgorithm::Blake2b256,
-							cid_codec: sp_transaction_storage_proof::runtime_api::RAW_CID_CODEC,
+							hashing: sp_transaction_storage_proof::HashingAlgorithm::Blake2b256,
+							cid_codec: sp_transaction_storage_proof::RAW_CID_CODEC,
 							extrinsic_index: u32::MAX,
-						}
 						})
 						.collect()
 				})
